first commit for july structured
