+++ .env
@@ -1,0 +1,1 @@
+MONGO_URL=mongodb+srv://Alpha_1:Alpha_1@ssms-server.irr0ltm.mongodb.net/?retryWrites=true&w=majority&appName=SSMS-Server
