--- mainDatabase.students.json
@@ -1,42 +1,0 @@
-[{
-  "_id": {
-    "$oid": "668cfa520f1de6cc6c7b0761"
-  },
-  "UserName": "raj",
-  "LastName": "roy",
-  "ParentName": "himadri roy",
-  "PhoneNumber": "1234567899",
-  "Email": "acas@da.com",
-  "Class": "8",
-  "Section": "c",
-  "SPOC": "",
-  "__v": 0
-},
-{
-  "_id": {
-    "$oid": "668d0f05473410cced07c382"
-  },
-  "UserName": "shreya",
-  "LastName": "debnath",
-  "ParentName": "Kamalesh Debnath",
-  "PhoneNumber": "12345678",
-  "Email": "shreya@gmail.com",
-  "Class": "5",
-  "Section": "D",
-  "SPOC": "",
-  "__v": 0
-},
-{
-  "_id": {
-    "$oid": "668d0f2d473410cced07c387"
-  },
-  "UserName": "sayani",
-  "LastName": "deb",
-  "ParentName": "Kamal Debnath",
-  "PhoneNumber": "12345678",
-  "Email": "shreya@gmsail.com",
-  "Class": "5",
-  "Section": "D",
-  "SPOC": "",
-  "__v": 0
-}]
