--- mainDatabase.superadmins.json
@@ -1,10 +1,0 @@
-[{
-  "_id": {
-    "$oid": "668cfad50f1de6cc6c7b0772"
-  },
-  "UserName": "user2",
-  "role": "superadmin",
-  "Email": "jane.smith2@example.com",
-  "PhoneNumber": "987-654-3210",
-  "__v": 0
-}]
