+++ models/user.js
@@ -1,0 +1,18 @@
+const mongoose = require("mongoose");
+
+// User schema
+const userSchema = new mongoose.Schema(
+  {
+    UserName: { type: String, required: true, unique: true },
+    FirstName: { type: String, required: true },
+    LastName: { type: String, required: true },
+    Password: { type: String, required: true },
+    Email: { type: String, required: true, unique: true },
+    PhoneNumber: { type: String, required: true },
+    Address: { type: String, required: true },
+    lastLogin: { type: Date, default: null },
+  },
+  { timestamps: true }
+);
+
+module.exports = mongoose.model("User", userSchema);
